+++ crates/cli/src/modules/mod.rs
@@ -55,7 +55,7 @@ pub struct Principal {
 
     #[serde(rename = "usedQuota")]
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub used_quota: Option<u32>,
+    pub used_quota: Option<u64>,
 
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub name: Option<String>,
@@ -145,7 +145,7 @@ pub enum PrincipalAction {
 pub enum PrincipalValue {
     String(String),
     StringList(Vec<String>),
-    Integer(u32),
+    Integer(u64),
 }
 
 impl PrincipalUpdate {
